+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8JXT2NX5C5P9RTEJ7R8MPSC"
+        "01K8MJYWC17ZECFNE9Z0NRX8E5"
     )
     r"""The unique identifier of the data source"""