+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -121,5 +121,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-27T13:30:23.405Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-28T04:59:15.145Z")
     r"""The date and time the resource was last updated"""